backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.2
motor>=3.3.2
pymongo>=4.6.0
celery>=5.3.4
redis>=5.0.1
google-generativeai>=0.3.2
pillow>=10.2.0
requests>=2.31.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
loguru>=0.7.2
